Added parameter checks and spec tests
The default values for computers_ou and users_ou has been changed from
dummy values to 'UNSET' to be able to more reliably check wheter they
are set or not. This means that the users_ou parameter is no longer
mandatory and can be ignored if RFC2307 mode is desired.

The general idea it that users_ou will be deprecated in favor of
upm_search_path, which better represents what the parameter is
controlling. This is to conform with parameters user_search_path and
group_search_path that is used to restrict the scope in RFC2307 mode.

Additionally control of the parameters root-update-mode and
group-update-mode has been added to be able to work around a problem
seen when using Workstation mode in RFC2307 mode. When this parameter is
set to none, which previously was hard coded in the vas.conf template,
users was sometimes not able to login. The default value is still 'none'