Update VAS API logic for API issues
Before this commit the function retrieving data from the VAS API would
return an empty array even though the API was down. This is the same
behaviour as if the API there are no entries for this host in the API.
This causes Puppet to remove all previously added entries from the API
if it happens to be down for whatever reason.

After this commit the function will instead return a status code and the
contents from the API (or error if it occurs). If the status code is anything
other than 200 occurs Puppet will send a warning() and not managed the
vas_users_allow resource.